Fixes for Gherkin 3.x.
Updating tests so that they are either compatible with or do not run
against version 3.x of the 'gherkin' gem. Because gherkin 2.x and 4.x
allow for a wider variety of input, gherkin 3.x is effectively a
subset of 2.x/4.x and, therefore, it should be safe to not run every
test against 3.x in addition to running them against 2.x/4.x.