Model output bug fix.
Fixed a bug that was causing example models to output extra newline
characters under certain circumstances. This behavior was hidden by a
method used in testing being called in a way that was ignoring the
extra newline characters. The method is now called in a way that does
not ignore extra newline characters.